refact: move room card component to more context path
